+++ railties/lib/rails/generators/rails/app/templates/config/initializers/application_controller_renderer.rb
@@ -1,0 +1,6 @@
+## Change renderer defaults here.
+#
+# ApplicationController.renderer.defaults.merge!(
+#   http_host: 'example.org',
+#   https: false
+# )